Review fix. moving to methods to repository
